requirements.txt
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.102.0
pillow==10.0.1
pypdf2==3.0.1
reportlab==3.6.0
